modified

+++ todo.js
@@ -1,14 +1,27 @@
+//JavaScript for noteMaker.html
+itemArrayStr = localStorage.getItem("itemsJSON");
+itemArray = JSON.parse(itemArrayStr);
+add = document.getElementById("addItem")
+add.addEventListener('click', update);
+display();
+clearList = document.getElementById("clearList");
+clearList.addEventListener("click", clearListFunc);
+
+// functions start from here
+
 function deleteItem(index) {
     console.log("calling delete...", index)
-    itemArrayStr = localStorage.getItem("itemsJSON");
-    itemArray = JSON.parse(itemArrayStr);
-    itemArray.splice(index, 1);
-    localStorage.setItem('itemsJSON', JSON.stringify(itemArray));
-    display();
+    let result = confirm("Want to delete this item  ?");
+    if (result) {
+        itemArray.splice(index, 1);
+        localStorage.setItem('itemsJSON', JSON.stringify(itemArray));
+        display();
+    }
+
 }
 function clearListFunc() {
-    itemArrayStr = localStorage.getItem("itemsJSON");
-    itemArray = JSON.parse(itemArrayStr);
+    // itemArrayStr = localStorage.getItem("itemsJSON");
+    // itemArray = JSON.parse(itemArrayStr);
     if (localStorage.getItem('itemsJSON') != null && itemArray.length > 0) {
         let result = confirm("Are you sure you want to delete?");
         if (result) {
@@ -32,8 +45,6 @@ function display() {
             <th scope="col">Action</th>
         </tr>`
     }
-    itemArrayStr = localStorage.getItem("itemsJSON");
-    itemArray = JSON.parse(itemArrayStr);
     let tableBody = document.getElementById("tableBody");
     let str = "";
     if (itemArray != null && itemArray.length > 0)
@@ -56,27 +67,22 @@ function update() {
     console.log("updating list....")
     title = document.getElementById("title").value;
     desc = document.getElementById("description").value;
-    if (localStorage.getItem("itemsJSON") == null) {
-        itemArray = [];
-        itemArray.unshift([title, desc]);
-        localStorage.setItem("itemsJSON", JSON.stringify(itemArray));
-        console.log(`added item with title "${title}" to list.`)
+    if (title != "") {
+        if (localStorage.getItem("itemsJSON") == null) {
+            itemArray = [];
+            itemArray.unshift([title, desc]);
+            localStorage.setItem("itemsJSON", JSON.stringify(itemArray));
+            console.log(`added item with title "${title}" to list.`)
+        } else {
+            itemArray.unshift([title, desc]);
+            localStorage.setItem("itemsJSON", JSON.stringify(itemArray));
+            console.log(`added item with title "${title}" to list.`);
+        }
     } else {
-        itemArrayStr = localStorage.getItem("itemsJSON");
-        itemArray = JSON.parse(itemArrayStr);
-        itemArray.unshift([title, desc]);
-        localStorage.setItem("itemsJSON", JSON.stringify(itemArray));
-        console.log(`added item with title "${title}" to list.`);
+        confirm('Title cannot be empty!!!')
     }
     display();
 }
-itemArrayStr = localStorage.getItem("itemsJSON");
-itemArray = JSON.parse(itemArrayStr);
-add = document.getElementById("addItem")
-add.addEventListener('click', update);
-display();
-clearList = document.getElementById("clearList");
-clearList.addEventListener("click", clearListFunc);
-// deleteItem = document.getElementById("deleteItem");
-// deleteItem.addEventListener('click', deleteItemFunc);
+
+
